Added ESP32 Client Code
This is Arduino code for the ESP32 which sets it as a station under your router and allows it to act as a client, connecting to a server of your choice and reading any data which may be transmitted through TCP. It will then echo that data back to the server as well as to any Serially connected devices. Details are commented on the code.