Added Vision Quartus Project to Main Project
IP modules from OpenCores and Qsys are included. Apologies for the large
amount of files, quite difficult to seperate what is necessaary and what
is unnecessary as there are literally too many files.

Current version has SPI implemented at a hardware level, but not at a
software level (NIOS II is incapable of using it)