Added RGB To HSV Conversion Module
Is it fast enough? Might need to pipeline the division so that we can
achieve a higher clock speed.

Threshold values will also need to be adjusted for testing condtions, at
the moment is hardware only, which causes long compilation times, but
adding the functionality for changing it from software shouldn't be that
difficult. Additionally, once we get full image sending out, we
shouldn't need to adjust the bounds manually, can just use the image
data for better accuracy.

Would like to change the colour detection to a HSV and neural net model,
but that's a different story.